LayoutTests/imported/w3c/web-platform-tests/dom/historical-expected.txt


FAIL Historical DOM features must be removed: CDATASection assert_equals: expected (undefined) undefined but got (object) object "[object CDATASectionConstructor]"
PASS Historical DOM features must be removed: DOMConfiguration 
PASS Historical DOM features must be removed: DOMError 
PASS Historical DOM features must be removed: DOMErrorHandler 
PASS Historical DOM features must be removed: DOMImplementationList 
PASS Historical DOM features must be removed: DOMImplementationSource 
PASS Historical DOM features must be removed: DOMLocator 
PASS Historical DOM features must be removed: DOMObject 
PASS Historical DOM features must be removed: DOMUserData 
PASS Historical DOM features must be removed: Entity 
PASS Historical DOM features must be removed: EntityReference 
PASS Historical DOM features must be removed: EventException 
PASS Historical DOM features must be removed: NameList 
PASS Historical DOM features must be removed: Notation 
PASS Historical DOM features must be removed: TypeInfo 
PASS Historical DOM features must be removed: UserDataHandler 
PASS Historical DOM features must be removed: RangeException 
FAIL Historical DOM features must be removed: createCDATASection assert_equals: expected (undefined) undefined but got (function) function "function createCDATASection() {
    [native code]
}"
PASS Historical DOM features must be removed: createEntityReference 
FAIL Historical DOM features must be removed: xmlEncoding assert_equals: expected (undefined) undefined but got (object) null
FAIL Historical DOM features must be removed: xmlStandalone assert_equals: expected (undefined) undefined but got (boolean) false
FAIL Historical DOM features must be removed: xmlVersion assert_equals: expected (undefined) undefined but got (object) null
PASS Historical DOM features must be removed: strictErrorChecking 
PASS Historical DOM features must be removed: domConfig 
PASS Historical DOM features must be removed: normalizeDocument 
PASS Historical DOM features must be removed: renameNode 
FAIL Historical DOM features must be removed: defaultCharset assert_equals: expected (undefined) undefined but got (string) "ISO-8859-1"
FAIL Historical DOM features must be removed: height assert_equals: expected (undefined) undefined but got (number) 600
FAIL Historical DOM features must be removed: width assert_equals: expected (undefined) undefined but got (number) 800
PASS DOMImplementation.getFeature() must be nuked. 
PASS Historical DOM features must be removed: schemaTypeInfo 
PASS Historical DOM features must be removed: setIdAttribute 
PASS Historical DOM features must be removed: setIdAttributeNS 
PASS Historical DOM features must be removed: setIdAttributeNode 
PASS DocumentType member must be nuked: entities 
PASS DocumentType member must be nuked: notations 
FAIL DocumentType member must be nuked: internalSubset assert_equals: expected (undefined) undefined but got (object) null
PASS Text member must be nuked: isElementContentWhitespace 
FAIL Text member must be nuked: replaceWholeText assert_equals: expected (undefined) undefined but got (function) function "function replaceWholeText() {
    [native code]
}"
PASS Node member must be nuked: hasAttributes 
PASS Node member must be nuked: attributes 
FAIL Node member must be nuked: namespaceURI assert_equals: expected (undefined) undefined but got (object) null
FAIL Node member must be nuked: prefix assert_equals: expected (undefined) undefined but got (object) null
FAIL Node member must be nuked: localName assert_equals: expected (undefined) undefined but got (object) null
FAIL Node member must be nuked: isSupported assert_equals: expected (undefined) undefined but got (function) function "function isSupported() {
    [native code]
}"
PASS Node member must be nuked: getFeature 
PASS Node member must be nuked: getUserData 
PASS Node member must be nuked: setUserData 
PASS Window member must be nuked: attachEvent 
